added comunicado controller

+++ src/main/java/com/xicola/xicola/repository/TipoPessoaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoPessoaRepository extends JpaRepository<TipoPessoa, Long> {
+public interface TipoPessoaRepository extends JpaRepository<TipoPessoa, Integer> {
 
 
 }
